requirements.txt
streamlit
pandas
plotly
websocket-client
scipy
statsmodels
scikit-learn
streamlit-autorefresh
